Fix to store milesValue in int

+++ src/main/java/RewardValue.java
@@ -1,11 +1,11 @@
 public class RewardValue {
     private double cashValue;
-    private double milesValue;
+    private int milesValue;
     private double milesToCashRate = 0.0035;
 
     public RewardValue(double cashValue) {
         this.cashValue = cashValue;
-        this.milesValue = cashValue / this.milesToCashRate;
+        this.milesValue = (int) (cashValue / this.milesToCashRate);
     }
 
     public RewardValue(int milesValue) {
